requirements.txt
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4
awesome-slugify==1.6.5
boto3==1.9.6
botocore==1.12.6
certifi==2018.8.24
chardet==3.0.4
click==6.7
docutils==0.14
Flask==1.0.2
flask-marshmallow==0.9.0
Flask-RESTful==0.3.6
Flask-WTF==0.14.2
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
lazy-object-proxy==1.3.1
MarkupSafe==1.0
marshmallow==2.15.5
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.7.1
py==1.6.0
pycodestyle==2.4.0
pylint==2.1.1
pylint-flask==0.5
pylint-plugin-utils==0.4
pytest==3.8.0
python-dateutil==2.7.3
pytz==2018.5
regex==2018.8.29
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
Unidecode==0.4.21
urllib3==1.23
webargs==4.1.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
